Conflicts:
	workspace/VS_glukas_AntiTheft/bin/VS_glukas_AntiTheft.apk
	workspace/VS_glukas_AntiTheft/bin/classes.dex

workspace/VS_glukas_Sensors/src/ethz/ch/vs/a1/glukas/sensors/SensorActiv
ity.java

+++ workspace/VS_glukas_AntiTheft/src/ch/ethz/inf/vs/a1/vs/glukas/antitheft/Settings.java
@@ -5,7 +5,7 @@ public class Settings {
 	public static final String SETTINGS_FILENAME = "AntiTheftSettings";
 
 	public static final boolean ACTIVATE_DEFAULT = false;
-	public static final int SENSITIVITY_DEFAULT = 0;
+	public static final double SENSITIVITY_DEFAULT = 0;
 	public static final int TIMEOUT_DEFAULT = 100;
 
 	public static final String ACTIVATE_STR = "activate";
